give permission to user controller for part of => index/store/update/delete/restore

+++ app/Http/Controllers/UserController.php
@@ -12,10 +12,14 @@ use Symfony\Component\HttpKernel\Profiler\Profile;
 
 class UserController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $users = User::orderBy('id', 'desc')->paginate(10);
-        return response()->json(['users' => $users]);
+        if ($request->user()->can('user.index')) {
+            $users = User::orderBy('id', 'desc')->paginate(10);
+            return response()->json(['users' => $users]);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
+        }
     }
 
     // public function store(StoreUserRequest $request)
@@ -25,7 +29,7 @@ class UserController extends Controller
     //     return response()->json(['message' => 'کاربر با موفقیت ایجاد شد', 'user' => $User], 201);
     // }
 
-    public function show($id)
+    public function show(Request $request, $id)
     {
         $user = User::with('tickets', 'comments')->findOrFail($id);
         return response()->json(['user' => $user]);
@@ -35,22 +39,34 @@ class UserController extends Controller
 
     public function update(UpdateUserRequest $request, string $id)
     {
-        $User = User::where('id', $id)->update($request->toArray());
-        return response()->json(['کاربر با موفقیت به روزرسانی شد', 'user' => $User]);
+        if ($request->user()->can()) {
+            $User = User::where('id', $id)->update($request->toArray());
+            return response()->json(['کاربر با موفقیت به روزرسانی شد', 'user' => $User]);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
+        }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
-        return response()->json(['message' => 'کاربر با موفقیت حذف شد.'], 200);
+        if ($request->user()->can('user.delete')) {
+            $user = User::findOrFail($id);
+            $user->delete();
+            return response()->json(['message' => 'کاربر با موفقیت حذف شد.'], 200);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
+        }
     }
 
-    public function restore($id)
+    public function restore(Request $request, $id)
     {
-        $user = User::onlyTrashed()->findOrFail($id);
-        $user->restore();
-        return response()->json(['message' => 'کاربر با موفقیت بازیابی شد.'], 200);
+        if ($request->user()->can('user.restore')) {
+            $user = User::onlyTrashed()->findOrFail($id);
+            $user->restore();
+            return response()->json(['message' => 'کاربر با موفقیت بازیابی شد.'], 200);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
+        }
     }
 
     public function  me()
@@ -62,13 +78,17 @@ class UserController extends Controller
 
     public function uploadProfileFile(Request $request, $id)
     {
-        $user = User::findOrFail($id);
+        if ($request->user()->can('user.uploadprofile')) {
+            $user = User::findOrFail($id);
 
-        if ($request->hasFile('profile')) {
-            $user->clearMediaCollection('profile');
-            $user->addMediaFromRequest('profile')->toMediaCollection('profile');
+            if ($request->hasFile('profile')) {
+                $user->clearMediaCollection('profile');
+                $user->addMediaFromRequest('profile')->toMediaCollection('profile');
+            }
+
+            return response()->json(['message' => 'پروفایل شما با موفقیت آپلود شد.']);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
         }
-
-        return response()->json(['message' => 'پروفایل شما با موفقیت آپلود شد.']);
     }
 }
